Download and resize image to store with item
When an imageUrl is provided, download the image, resize it
to 40x40 and store the raw image with the item object so
that the browser doesn't need to get all the images separately.

This should increase the load speed of the dashboard when there
are a lot of items. The images are stored at 40x40 as that is the
currently configured image size being rendered in the client