Loop with Grid component
Previously the loop was inside a Grid item component
this was causing all of the itemCards to be in a single
Grid component so it was not formatting as expected.
Updated the loop to create separate Grid items per
itemCard componenets. Items are now flexing properly
however, still need to center vertically