Add item lookup on POST and PUT
This will query the newly added or updated item in the database
and return it in the response. The UI can then update the page
with the new values.

The query is necessary due to the image processing that happens
preventing just returning the object that was inserted or updated
without the additional query