Auto update item list on change (#22)
* Add item lookup on POST and PUT

This will query the newly added or updated item in the database
and return it in the response. The UI can then update the page
with the new values.

The query is necessary due to the image processing that happens
preventing just returning the object that was inserted or updated
without the additional query

* Add prop to editDialog to update parent state

Edit functionality updates with parent state in itemGrid
Adding item not yet in place

* Reload page when adding new item

This is more of a workaround since the itemList is retrieved in the onload
event and the add item button is in the navbar.