Client config using env vars and window object
Using environment variables for changing config values without
requiring a rebuild and deployment of the client code.
WIP need to adjust for use while running outside of docker.
Need a way to set the window.ENV without the docker-entrypoint
script so that the standard npm start can be use during development