Remove dotenv and use only as dev dependency
Exclude dotenv from code as this is expected to run as a container
using environments in a container is expected over a file. Locally,
setting the environment via file is easier.
Added additional npm script to run the server without the dotenv flag
when running as a deployed application/container