Add authentication feature for sign in with new dialog WIP
Added launch.json file for debugging
The sign in is not yet tested fully. Only worked with the
auth disabled so far. Sometimes, the response does not return at all
from the API. Not sure why, but need to test with auth enabled to be
sure it's not a DB problem since there is no users database.