Update API settings auth disabled check to use equivalency
The previous version used a null coalesce which resulted in the settings being
returned to the UI as a string. This led to the UI if checks passing even when
the value was false due to the javascript 'truthy' checks.

This caused problems when a user attempted to perform and admin function as the
API would reject it, and the UI did not provide the login interface