Add seeder script and additional local npm scripts to aid development
Added a nodejs script to generate items and insert them into the local
MongoDB instance to aid in testingand debugging without the need to
manually create items. This allows an unlimited amount of items to be
inserted into the local database.
New local: npm scripts to facilitate running the application with the
local database regardless of the .env configuration to allow easy
transition from personal items on a separate database to debugging
with the local MongoDB instance